requirements.txt
# This list of dependencies was generated with `pip3 freeze`

# Logging and parse args
absl-py==1.0.0

# Gym environment
gym==0.23.1
ale-py==0.7.4
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
box2d-py==2.3.8
pyglet==1.5.21
pygame==2.1.2

# PyTorch
torch==2.2.2
torchsummary==1.5.1
torchvision==0.17.2

# Utilities
numpy==1.21.6
opencv-python==4.5.5.62
python-snappy==0.6.0

# Coding styles and others
pre-commit==2.17.0
black==22.1.0

# transformers
transformers==4.40.2
open-clip-torch==2.24.0
omegaconf==2.3.0
accelerate==0.30.0
boto3
cached_path
botocore
rich
